+++ src/core/interfaces/IOutswapV1Callee.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.24;
 
 interface IOutswapV1Callee {
-    function OutswapV1Call(address sender, uint amount0, uint amount1, bytes calldata data) external;
+    function OutswapV1Call(address sender, uint256 amount0, uint256 amount1, bytes calldata data) external;
 }
