+++ src/libraries/OutswapV1OracleLibrary.sol
@@ -1,8 +1,8 @@
 //SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.24;
 
-import './FixedPoint.sol';
-import '../core/interfaces/IOutswapV1Pair.sol';
+import "./FixedPoint.sol";
+import "../core/interfaces/IOutswapV1Pair.sol";
 
 // library with helper methods for oracles that are concerned with computing average prices
 library OutswapV1OracleLibrary {
@@ -14,9 +14,11 @@ library OutswapV1OracleLibrary {
     }
 
     // produces the cumulative price using counterfactuals to save gas and avoid a call to sync.
-    function currentCumulativePrices(
-        address pair
-    ) internal view returns (uint price0Cumulative, uint price1Cumulative, uint32 blockTimestamp) {
+    function currentCumulativePrices(address pair)
+        internal
+        view
+        returns (uint256 price0Cumulative, uint256 price1Cumulative, uint32 blockTimestamp)
+    {
         blockTimestamp = currentBlockTimestamp();
         price0Cumulative = IOutswapV1Pair(pair).price0CumulativeLast();
         price1Cumulative = IOutswapV1Pair(pair).price1CumulativeLast();
@@ -28,9 +30,9 @@ library OutswapV1OracleLibrary {
             uint32 timeElapsed = blockTimestamp - blockTimestampLast;
             // addition overflow is desired
             // counterfactual
-            price0Cumulative += uint(FixedPoint.fraction(reserve1, reserve0)._x) * timeElapsed;
+            price0Cumulative += uint256(FixedPoint.fraction(reserve1, reserve0)._x) * timeElapsed;
             // counterfactual
-            price1Cumulative += uint(FixedPoint.fraction(reserve0, reserve1)._x) * timeElapsed;
+            price1Cumulative += uint256(FixedPoint.fraction(reserve0, reserve1)._x) * timeElapsed;
         }
     }
 }
